+++ backend/database/migrations/2026_06_28_145532_create_meeting_minutes_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->unsignedInteger('meeting_id');
             $table->longText('discussion_summary')->nullable();
             $table->enum('status', ['draft', 'pending_approval', 'approved'])->default('draft');
-            $table->unsignedInteger('created_by');
+            $table->unsignedBigInteger('created_by');
             $table->timestamp('created_at')->useCurrent();
             $table->timestamp('updated_at')->useCurrent()->useCurrentOnUpdate();
